refactor: move MCP resolution into runtime client
